% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{AIC.dynsbm}
\alias{AIC.dynsbm}
\title{AIC method for DynSBM Fits}
\usage{
\method{AIC}{dynsbm}(x)
}
\arguments{
\item{x}{an object of class "dynsbm"}
}
\value{
returns the Akaike Information Criterion using the posterior mean
parameters as approximations to the MLE.  The number of parameters is taken
to be T*(2*(n-k) + k*k) + n.
}
\description{
Returns Akaike Information Criterion for DynSBM Fit
}
\seealso{
\code{\link{dynsbm}}
}

