% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{AIC.wsbm}
\alias{AIC.wsbm}
\title{AIC method for WSBM Fits}
\usage{
\method{AIC}{wsbm}(x)
}
\arguments{
\item{x}{an object of class "wsbm"}
}
\value{
returns the Akaike Information Criterion using the posterior mean
parameters as approximations to the MLE.  The number of parameters is taken
to be 2*(n-k) + k*k + n.
}
\description{
Returns Akaike Information Criterion for WSBM Fit
}
\seealso{
\code{\link{wsbm}}
}

