% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{BIC.dynsbm}
\alias{BIC.dynsbm}
\title{BIC method for DynSBM Fits}
\usage{
\method{BIC}{dynsbm}(object, ...)
}
\arguments{
\item{object}{an object of class "dynsbm"}

\item{...}{additional arguments to BIC}
}
\value{
returns the Bayesian Information Criterion using the posterior mean
parameters as approximations to the MLE.  The number of parameters is taken
to be T*(2*(n-k) + k*k) + n and the sample size is taken to be T*n*n.
}
\description{
Returns Bayesian Information Criterion for DynBM Fit
}
\seealso{
\code{\link{dynsbm}}
}
