% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{BIC.wsbm}
\alias{BIC.wsbm}
\title{BIC method for WSBM Fits}
\usage{
\method{BIC}{wsbm}(x)
}
\arguments{
\item{x}{an object of class "wsbm"}
}
\value{
returns the Bayesian Information Criterion using the posterior mean
parameters as approximations to the MLE.  The number of parameters is taken
to be 2*(n-k) + k*k + n and the sample size is taken to be n*n.
}
\description{
Returns Bayesian Information Criterion for WSBM Fit
}
\seealso{
\code{\link{wsbm}}
}

