% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{block.prior.plot.dynsbm.mcmc}
\alias{block.prior.plot.dynsbm.mcmc}
\title{Block Prior Plotting Method for DynSBM Class}
\usage{
block.prior.plot.dynsbm.mcmc(object, remap = NULL, truth = NULL,
  blocks = NULL, ...)
}
\arguments{
\item{object}{object of class "dynsbm"}

\item{remap}{relabeling of blocks that is a permuation of 1:kk}

\item{truth}{true values for comparison if known}

\item{blocks}{subset of the blocks 1:kk for which to plot posteriors}

\item{...}{additional parameters to pass to plotting functions}
}
\description{
Method for plotting posterior distribution of block prior means across
  DynSBM classes
}
