% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{commuter30}
\alias{commuter30}
\title{Simulated Commuter Network}
\format{An object of class "wsbm"}
\usage{
commuter30
}
\description{
A dataset containing the model parameters and simulated network
from a weighted stochastic block model.
}
\keyword{datasets}
