% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{create.wsbm.list}
\alias{create.wsbm.list}
\title{Convert DynSBM Object to List of WSBM Objects}
\usage{
create.wsbm.list(object)
}
\arguments{
\item{object}{object of class "dynsbm"}
}
\value{
Returns a list of length TT where each object in the list is of
  class "wsbm".
}
\description{
Convert DynSBM Object to List of WSBM Objects
}
