% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{data.gen.wsbm}
\alias{data.gen.wsbm}
\title{Generate Weighted Stochastic Block Model Networks}
\usage{
data.gen.wsbm(nn = 50, BB = array(c(5, 1, 1, 5), c(2, 2)),
  mmb.prior = c(0.5, 0.5), mmb, SS = rep(1, nn), RR = rep(1, nn),
  self.ties = TRUE, hours = 1)
}
\arguments{
\item{nn}{Number of nodes in the network}

\item{BB}{Block propensity matrix}

\item{mmb.prior}{Prior distribution for multinomial distribution over
block memberships}

\item{mmb}{Block memberhsip vector}

\item{SS}{Sender effect vector}

\item{RR}{Receiver effect vector}

\item{self.ties}{If true self ties are allowed}

\item{hours}{Scaling factor so that other factors can be interpreted
as hourly rate parameters.}
}
\value{
Returns a list containing the simulated network as the first
component 'net'.  The list also contains all of the parameters used to
simulate the network
}
\description{
Generates networks from the weighted stochastic block model
with or without degree effects.
}
\examples{
kk <- 3
nn <- kk^2 * 10
mmb <-  rep(1:kk,each=nn/kk)
BB <- array(1,c(kk,kk)); diag(BB) <- 5
SS <- RR <- rep(c(rep(5,nn/(kk^2)),rep(1,nn*(kk-1)/(kk^2))),kk)
set.seed(100)
dat <- data.gen.wsbm(nn=nn,mmb=mmb,BB=BB,SS=SS,RR=RR)

}
\seealso{
\code{\link{wsbm}}
}

