% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dc60}
\alias{dc60}
\title{Simulated Dynamic Commuter Network}
\format{An object of class "dynsbm":
\describe{
  \item{TT=10}{Number of networks in the dataset}
  \item{ee=2}{Number of equivalence classes}
  \item{nn=60}{Number of nodes}
  \item{mmb}{The network contains 3 communities of equal size}
}}
\usage{
dc60
}
\description{
A dataset containing the model parameters and simulated networks
from a dynamic weighted stochastic block model.
}
\keyword{datasets}
