% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.prior.R
\name{dgamma.prior}
\alias{dgamma.prior}
\alias{rgamma.prior}
\title{The Gamma Conjugate Prior Distribution}
\usage{
dgamma.prior(x, p = 0.99, q = 1.01, r = 1, s = r, log = FALSE)

rgamma.prior(n, p = 0.99, q = 1.01, r = 1, s = r, ...)
}
\arguments{
\item{x}{vector of length 2 at which to evaluate density}

\item{p}{parameter representing product of samples from a Gamma distribution}

\item{q}{parameter representing sum of samples from a Gamma distribution}

\item{r}{parameter representing samples used to calculate product}

\item{s}{parameter representing samples used to calculate sum}

\item{log}{if TRUE, probabilities P are returned as log(P)}

\item{n}{number of samples to draw}

\item{...}{additional parameters for conjugate Gamma prior sampler}
}
\description{
Density and random generation for the conjugate prior to the Gamma
  distribution.
}
\details{
Returns the density for points drawn from the bivariate conjugate
  gamma prior.
}
