% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{diagnostic.plot.dynsbm.mcmc}
\alias{diagnostic.plot.dynsbm.mcmc}
\title{Diagnostic Plotting Method for DynSBM fits}
\usage{
\method{diagnostic.plot}{dynsbm.mcmc}(x, t, marginal = TRUE,
  param = c("block", "sender", "receiver", "ll"), ...)
}
\arguments{
\item{x}{object of class "dynsbm"}

\item{t}{network or class number to plot diagnostics for}

\item{marginal}{if TRUE, plots diagnostics for class level parameters,
otherwise plots diagnostics for network level parameters}

\item{param}{which parameters diagnostics to plot.  The default value "block"
plots the block intensity parameters.  Passing "sender" and "receiver" plots
the chains for the sender and receiver effects.  The last option, "ll",
plots the Log-Likelihood for each step of the chain.}

\item{...}{other graphical parameters to pass to the plotting functions}

\item{scale.ylim}{logical indicating whether each chain should have the same
range for the y-axis.  This can be suppressed by passing ylim.}
}
\description{
Plots MCMC chains for various parameters in WSBM model.
}
\details{
Plots MCMC chains for the specified set of parameters.
}
\examples{
data(commuter30)
fit.wsbm <- wsbm(commuter30,kk=3,hours=1)

diagnostic.plot(fit.wsbm,param="block")
diagnostic.plot(fit.wsbm,param="sender")
diagnostic.plot(fit.wsbm,param="receiver")
diagnostic.plot(fit.wsbm,param="ll")

}
\seealso{
\code{\link{wsbm}}
}

