% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/wsbm.R, R/dynamic-sbm.R
\name{diagnostic.plot}
\alias{diagnostic.plot}
\alias{diagnostic.plot.wsbm.mcmc}
\alias{diagnostic.plot.dynsbm.mcmc}
\title{Diagnostic Plotting Function}
\usage{
diagnostic.plot(object, ...)

\method{diagnostic.plot}{wsbm.mcmc}(object, param = c("block", "sender",
  "receiver", "ll"), scale.ylim = FALSE, ...)

\method{diagnostic.plot}{dynsbm.mcmc}(object, marginal = TRUE, t,
  param = c("block", "sender", "receiver", "ll"), ...)
}
\arguments{
\item{object}{model object}

\item{...}{additional arguments for plotting methods}

\item{param}{which diagnostics to plot.  The default value, "block",
plots the block intensity parameters.  Passing "sender" and "receiver" plots
the chains for the sender and receiver effects.  The last option, "ll",
plots the Log-Likelihood for each step of the chain.}

\item{scale.ylim}{logical indicating whether each chain should have the same
range for the y-axis.  This can be suppressed by passing ylim.}

\item{marginal}{if TRUE, plots diagnostics for class level parameters,
otherwise plots diagnostics for network level parameters}

\item{t}{if marginal = TRUE, t is the class number for which to plot
diagnostics.  If marginal = FALSE, t is the network number.}
}
\description{
Plots Diagnostics for Various Model Fits
}
\details{
Plots diagnostics for various parameter estimations for a given
  model object.
}
\section{Methods (by class)}{
\itemize{
\item \code{wsbm.mcmc}: Plots MCMC chains for various parameters
in WSBM model.

\item \code{dynsbm.mcmc}: Plots MCMC chains for various
parameters in DynSBM model.
}}

\examples{
### Static SBM Example
data(commuter30)
fit.1 <- wsbm.fit(commuter30$net,kk=3,hours=1)

diagnostic.plot(fit.1,param="block")
diagnostic.plot(fit.1,param="sender")
diagnostic.plot(fit.1,param="receiver")
diagnostic.plot(fit.1,param="ll")

### Dynamic SBM Example
data(dc60)
fit.dc <- dynsbm.fit(dc60$net.mat,kk=3,
                     tmap=dc60$tmap,hours.vec=dc60$hours.vec)

diagnostic.plot(fit.dc,t=1,param="block")
diagnostic.plot(fit.dc,t=2,param="block")

diagnostic.plot(fit.dc,t=1,param="sender")
diagnostic.plot(fit.dc,t=1,param="receiver")
diagnostic.plot(fit.dc,t=1,param="ll")

diagnostic.plot(fit.dc,t=1,param="block",marginal=FALSE)
}
\seealso{
\code{\link{wsbm.fit}}, \code{\link{dynsbm.fit}}
}
