% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{diagnostic.plot.wsbm.mcmc}
\alias{diagnostic.plot.wsbm.mcmc}
\title{Diagnostic Plotting Method for WSBM fits}
\usage{
\method{diagnostic.plot}{wsbm.mcmc}(x, param = c("block", "sender",
  "receiver", "ll"), scale.ylim = FALSE, ...)
}
\arguments{
\item{x}{object of class "wsbm"}

\item{scale.ylim}{logical indicating whether each chain should have the same
range for the y-axis.  This can be suppressed by passing ylim.}

\item{...}{other graphical parameters to pass to the plotting functions}

\item{type}{which parameters diagnostics to plot.  The default value "block"
plots the block intensity parameters.  Passing "sender" and "receiver" plots
the chains for the sender and receiver effects.  The last option, "ll",
plots the Log-Likelihood for each step of the chain.}
}
\description{
Plots MCMC chains for various parameters in WSBM model.
}
\details{
Plots MCMC chains for the specified set of parameters.
}
\examples{
data(commuter30)
fit.1 <- wsbm.fit(commuter30,kk=3,hours=1)

diagnostic.plot(fit.1,param="block")
diagnostic.plot(fit.1,param="sender")
diagnostic.plot(fit.1,param="receiver")
diagnostic.plot(fit.1,param="ll")

}
\seealso{
\code{\link{wsbm.fit}}
}

