% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{DIC}
\alias{DIC}
\title{Calculate Deviance Information Criterion}
\usage{
DIC(x, ...)
}
\arguments{
\item{x}{object for which to calculate DIC}

\item{...}{additional parameters}
}
\value{
Returns the Deviance Information Criterion.
}
\description{
\code{DIC} returns the Deviance Information Criterion for a Model Object
}
\details{
When comparing models using DIC, the smaller the value, the better
the fit.

This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the first
argument.
}
\seealso{
\code{\link{AIC}} \code{\link{BIC}}
}

