% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{DIC.wsbm}
\alias{DIC.wsbm}
\title{DIC method for WSBM Fits}
\usage{
\method{DIC}{wsbm}(x)
}
\arguments{
\item{x}{an object of class "wsbm"}
}
\value{
returns the Deviance Information Criterion.
}
\description{
Returns Deviance Information Criterion for WSBM Fit
}
\seealso{
\code{\link{wsbm}}
}

