% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{dynsbm.fit}
\alias{dynsbm.fit}
\title{Dynamic Weighted Stochastic Block Model Estimator}
\usage{
dynsbm.fit(net.mat, kk = 3, tmap, hours.vec, self.ties = TRUE,
  priors = dynsbm.priors(eta = rep(1/kk, kk)),
  init.control = list(spectral.start = FALSE, multistart = 0, multistart.total
  = 10), mcmc.control = list(total = 1000, burn.in = 1000, thin = 10,
  extend.alpha = 0.001, extend.max = 10, label.switch.mode = "adhoc",
  label.switch.max = 200, multi.impute = FALSE), update.mmb = TRUE,
  clean.out = FALSE, verbose = 1)
}
\arguments{
\item{net.mat}{a 3-dimensional array containing a sequence of adjacency
matrices.}

\item{kk}{number of blocks}

\item{tmap}{mapping between networks and equivalence classes}

\item{hours.vec}{vector of scaling parameters for each equivalence class}

\item{self.ties}{if true assumes self ties are possible}

\item{priors}{priors for posterior inference.  See dynsbm.priors for more
details}

\item{init.control}{control parameters for initialization routine.
If init.control contains a member named init.vals it is interpreted as a
list of parameters from which to start the MCMC chain.}

\item{mcmc.control}{control parameters for the MCMC algorithm}

\item{clean.out}{if true, removes MCMC chain from output and only
returns posterior means}

\item{verbose}{higher values correspond to more informative output as the
sampler runs.  Setting verbose = 0 suppresses all output.}
}
\value{
Returns a dynsbm object that has posterior means for each parameter
in the dynamic, weighted, directed,  degree corrected stochastic block model.
If clean.out is FALSE, this object also contains an element 'chain' which
contains the thinned and burned in draws from the MCMC sampler.  The prior
parameters from the call to dynsbm are also included.
}
\description{
Estimates parameters for the dynamic weighted stochastic block model using
an MCMC algorithm.
}
\seealso{
\code{\link{dynsbm.priors}}
}

