% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{dynsbm.priors}
\alias{dynsbm.priors}
\title{Prior Parameters for dynsbm}
\usage{
dynsbm.priors(eta = NULL, BB.hyperprior = c(0.9999, 0.0011, 0.001, 0.001),
  SS.hyperprior = c(0.999, 0.011, 0.01, 0.01), RR.hyperprior = c(0.999,
  0.011, 0.01, 0.01))
}
\arguments{
\item{eta}{prior on block memberships.}

\item{BB.hyperprior}{hyperprior parameters for conjugate Gamma
distribution on block effect priors}

\item{SS.hyperprior}{hyperprior parameters for conjugate Gamma
distribution on sender effect priors}

\item{RR.hyperprior}{hyperprior parameters for conjugate Gamma
distribution on receiver effect priors}
}
\value{
Returns a list containing the parameters with useful defaults.  eta
is the only required parameter.
}
\description{
Generating prior parameters for pass to dynsbm
}
\seealso{
\code{\link{dynsbm}}
}
