% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{dynsbm}
\alias{dynsbm}
\title{Create Dynamic Stochastic Block Model Object}
\usage{
dynsbm(nn = 40, TT = 4, tmap = rep(1:2, 2), hours.vec = c(1, 1),
  BB.prior = NULL, SS.prior = NULL, RR.prior = NULL, BB.mat, SS.mat,
  RR.mat, mmb.prior = c(0.5, 0.5), mmb, self.ties = TRUE,
  normalize = FALSE, gen = FALSE)
}
\arguments{
\item{nn}{number of nodes in the network}

\item{TT}{number of networks to generate}

\item{tmap}{mapping from time points 1:TT to equivalence classes}

\item{hours.vec}{number of hours for each equivalence class}

\item{BB.prior}{block effect prior parameters for each equivalence class}

\item{SS.prior}{sender effect prior parameters for each equivalence class}

\item{RR.prior}{receiver effect prior parameters for each equivalence class}

\item{mmb.prior}{prior distribution for multinomial distribution over
block memberships}

\item{mmb}{block memberhsip vector}

\item{self.ties}{if TRUE self ties are allowed}

\item{normalize}{if TRUE, sender/receiver effects are normalized to have a
mean of 1 within each block}

\item{gen}{if TRUE also samples a network from the model and includes it as a
component of the model object}
}
\value{
Returns a "dynsbm" model object containing all the parameters needed
for sampling networks from the model.  If gen = TRUE, the object also
contains a network sampled from this parameterization.
}
\description{
Generates a Model Object for a Dynamic Stochastic Block Model
}
\examples{
kk <- 3
nn <- kk^2 * 10
TT <- 10; ee <- 2
tmap <- rep(1:ee,TT/ee)
hours.vec <- c(12,12)
mmb <-  rep(1:kk,each=nn/kk)
BB.prior <- array(1,c(kk,kk,2,ee))
BB.prior[cbind(1:3,c(2,3,1),1,1)] <- 2
BB.prior[cbind(1:3,c(3,1,2),1,2)] <- 2

SS.prior <- array(rgamma(nn*2*ee,10,10),c(nn,2,ee))
RR.prior <- array(rgamma(nn*2*ee,10,10),c(nn,2,ee))
dat <- dynsbm(nn=nn,TT=TT,tmap=tmap,hours.vec=hours.vec,
              BB.prior=BB.prior,SS.prior=SS.prior,RR.prior=RR.prior
              mmb=mmb,self.ties=TRUE,normalize=TRUE,gen=TRUE)

}
\seealso{
\code{\link{dynsbm}}
}

