% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get.iter}
\alias{get.iter}
\title{Get Model Object for an MCMC Iteration}
\usage{
get.iter(x, iter, ...)
}
\arguments{
\item{x}{mdl.mcmc object}

\item{iter}{iteration to return}

\item{...}{}
}
\description{
Returns Model Object from MCMC Iteration
}
\details{
This is a generic function.  Given an object of type
}

