% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{logLik.dynsbm}
\alias{logLik.dynsbm}
\title{Log-Likelihood method for DynSBM Fits}
\usage{
\method{logLik}{dynsbm}(object, ...)
}
\arguments{
\item{object}{an object of class "dynsbm"}

\item{...}{additional arguments to logLik}
}
\value{
returns the log-likelihood of the data given the posterior mean
parameter estimates.
}
\description{
Returns log-likelihood for posterior mean estimates.
}
\seealso{
\code{\link{dynsbm}}
}
