% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvnm.R
\docType{package}
\name{lvnm}
\alias{lvnm}
\alias{lvnm-package}
\title{lvnm:  A package for estimating parameters in latent variable network
models}
\description{
Currently the lvnm package provides support for estimating weighted,
directed stochastic block models as well as the dynamic weighted
stochastic block model.
}
