% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{net.gen.dynsbm}
\alias{net.gen.dynsbm}
\title{Generate Set of Dynamic Stochastic Block Model Networks}
\usage{
\method{net.gen}{dynsbm}(x, marginal = FALSE)
}
\arguments{
\item{nn}{number of nodes in the network}

\item{TT}{number of networks to generate}

\item{tmap}{mapping from time points 1:TT to equivalence classes}

\item{hours.vec}{number of hours for each equivalence class}

\item{BB.prior}{block effect prior parameters for each equivalence class}

\item{SS.prior}{sender effect prior parameters for each equivalence class}

\item{RR.prior}{receiver effect prior parameters for each equivalence class}

\item{mmb.prior}{prior distribution for multinomial distribution over
block memberships}

\item{mmb}{block memberhsip vector}

\item{self.ties}{if TRUE self ties are allowed}

\item{normalize}{if TRUE, sender/receiver effects are normalized to have a
mean of 1 within each block}
}
\value{
Returns a list containing the simulated matrix of networks as the
first component 'net.mat'. The list also contains all of the parameters
used to simulate the networks.
}
\description{
Generates networks from the weighted stochastic block model
with or without degree effects.
}
\examples{
kk <- 3
nn <- kk^2 * 10
TT <- 10; ee <- 2
tmap <- rep(1:ee,TT/ee)
hours.vec <- c(12,12)
mmb <-  rep(1:kk,each=nn/kk)
BB.prior <- array(1,c(kk,kk,2,ee))
BB.prior[cbind(1:3,c(2,3,1),1,1)] <- 2
BB.prior[cbind(1:3,c(3,1,2),1,2)] <- 2

SS.prior <- array(rgamma(nn*2*ee,10,10),c(nn,2,ee))
RR.prior <- array(rgamma(nn*2*ee,10,10),c(nn,2,ee))
dat <- data.gen.dynsbm(nn=nn,TT=TT,tmap=tmap,hours.vec=hours.vec,
                       BB.prior=BB.prior,SS.prior=SS.prior,RR.prior=RR.prior
                       mmb=mmb,self.ties=TRUE,normalize=TRUE)

}
\seealso{
\code{\link{dynsbm}}
}

