% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{net.gen.matrix}
\alias{net.gen.matrix}
\title{Generate Adjacency Matrix}
\usage{
\method{net.gen}{matrix}(x, self.ties = TRUE, d = c("poisson", "bernoulli"))
}
\arguments{
\item{x}{object of class matrix}

\item{self.ties}{if FALSE, the diagonals are structurally missing}

\item{d}{distribution for sampling}
}
\value{
Returns a matrix representing the adjacency matrix for a network
drawn with the given tie intensities/probabilities
}
\description{
Generates an adjacency matrix given a tie intensity or tie probability
matrix
}
\examples{
nn <- 50
pmat <- array(5,c(nn,nn))
net.weighted <- net.gen(pmat)

pmat <- array(.25,c(nn,nn))
net.binary <- net.gen(pmat,d="b")

}
\seealso{
\code{\link{net.gen.wsbm}}, \code{\link{net.gen.dynsbm}}
}

