% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{net.gen}
\alias{net.gen}
\title{Network Generation Function}
\usage{
net.gen(object, ...)
}
\arguments{
\item{object}{model object}

\item{...}{additional arguments for specific methods}
}
\description{
Generates Adjacency Matrix Corresponding to a Model Object
}
\details{
This is a generic funciton.  See specific method definitions for
additional information.
}
\seealso{
\code{\link{net.gen.wsbm}},
\code{\link{net.gen.dynsbm}}
}
