% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{net.gen.wsbm}
\alias{net.gen.wsbm}
\title{Generate Weighted Stochastic Block Model Networks}
\usage{
\method{net.gen}{wsbm}(object, ...)
}
\arguments{
\item{object}{object of class "wsbm"}

\item{...}{additional parameters}
}
\value{
Returns a matrix representing the adjacency matrix for a network
drawn from the given weighted stochastic block model
}
\description{
Generates networks from the weighted stochastic block model
with or without degree effects.
}
\examples{
kk <- 3
nn <- kk^2 * 10
mmb <-  rep(1:kk,each=nn/kk)
BB <- array(1,c(kk,kk)); diag(BB) <- 5
SS <- RR <- rep(c(rep(5,nn/(kk^2)),rep(1,nn*(kk-1)/(kk^2))),kk)
set.seed(100)
model <- wsbm(nn=nn,mmb=mmb,BB=BB,SS=SS,RR=RR)
net <- net.gen(model)

plot(net)

par(mfrow=c(2,2))
for(ii in 1:4) plot(net.gen(model))


}
\seealso{
\code{\link{wsbm}} \code{\link{wsbm.fit}}
}
