% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{network.plot}
\alias{network.plot}
\title{Network Plot}
\usage{
network.plot(x, ...)
}
\arguments{
\item{x}{network object}

\item{...}{}
}
\description{
Plots a Heatmap Representation of an Adjacency Matrix
}
\details{
This is a generic function.
}
\seealso{
\code{\link{network.plot.wsbm}},
\code{\link{network.plot.dynsbm}}
}

