% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{param.plot.dynsbm.mcmc}
\alias{param.plot.dynsbm.mcmc}
\title{Parameter Plotting Method for DynSBM fits}
\usage{
\method{param.plot}{dynsbm.mcmc}(x, tclass, param = c("block", "sender",
  "receiver"), ...)
}
\arguments{
\item{x}{object of class "dynsbm"}

\item{param}{which parameters diagnostics to plot.  The default value "block"
plots the block intensity parameters.  Passing "sender" and "receiver" plots
the chains for the sender and receiver effects.  The last option, "ll",
plots the Log-Likelihood for each step of the chain.}

\item{...}{other graphical parameters to pass to the plotting functions}
}
\description{
Plots boxplots of posterior samples for parameters in DYNSBM model.
}
\details{
Plots MCMC chains for the specified set of parameters.
}
\examples{
data(commuter30)
fit.dynsbm <- dynsbm.fit(commuter30,kk=3,hours=1)

param.plot(fit.dynsbm,param="block")
param.plot(fit.dynsbm,param="sender")
param.plot(fit.dynsbm,param="receiver")
param.plot(fit.dynsbm,param="ll")

}
\seealso{
\code{\link{dynsbm.fit}}
}

