% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/wsbm.R, R/dynamic-sbm.R
\name{param.plot}
\alias{param.plot}
\alias{param.plot.wsbm.mcmc}
\alias{param.plot.dynsbm.mcmc}
\title{Parameter Plotting Function}
\usage{
param.plot(object, ...)

\method{param.plot}{wsbm.mcmc}(object, param = c("block", "sender",
  "receiver"), ...)

\method{param.plot}{dynsbm.mcmc}(object, tclass, param = c("block", "sender",
  "receiver"), ...)
}
\arguments{
\item{object}{model object
plots the Log-Likelihood for each step of the chain.}

\item{...}{other graphical parameters to pass to the plotting functions}

\item{param}{which parameters to use.  The default value "block"
plots the block intensity parameters.  Passing "sender" and "receiver" plots
the chains for the sender and receiver effects.}

\item{tclass}{class for which to plot parameters.}
}
\description{
Plots Parameter Summaries for Various Model Fits
}
\details{
Displays boxplots for various sets of parameters for the model
  object.
}
\section{Methods (by class)}{
\itemize{
\item \code{wsbm.mcmc}: Displays boxplots of the posterior sample for
parameters in the WSBM model.

\item \code{dynsbm.mcmc}: Displays boxplots of the posterior samples for
parameters within one of the equivalence classes specified by tclass.
}}

\examples{
###  WSBM Example
data(commuter30)
fit.1 <- wsbm.fit(commuter30$net,kk=3,hours=1)

param.plot(fit.1,param="block")
param.plot(fit.1,param="sender")
param.plot(fit.1,param="receiver")

}
\seealso{
\code{\link{wsbm.fit}}, \code{\link{dynsbm.fit}}
}
