% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{param.post.predict.dynsbm.mcmc}
\alias{param.post.predict.dynsbm.mcmc}
\title{Posterior Prediction of Future Parameter Values}
\usage{
\method{param.post.predict}{dynsbm.mcmc}(object, tclass, param = c("block",
  "sender", "receiver"), ...)
}
\arguments{
\item{object}{a fitted object of the class "dynsbm".  Requires
clean.out = FALSE
in the call to dynsbm.}

\item{tclass}{which equivalence class to compute posterior predictive
distribution for.}

\item{param}{which parameter to compute posterior predictive distribution
for.}

\item{...}{additional parameters}
}
\value{
Returns an array containing posterior predictive samples for the
given parameters.
}
\description{
Generates a posterior predictive distribution for parameters in the
DynSBM Model
}
\seealso{
\code{\link{dynsbm}} \code{\link{predict.dynsbm}}
}
