% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param.post.predict}
\alias{param.post.predict}
\title{Posterior Predictive Distribution Estimator}
\usage{
param.post.predict(object, ...)
}
\arguments{
\item{object}{model object which contains output from an MCMC chain}

\item{...}{additional arguments for specific methods}
}
\description{
Estimates Posterior Predictive Distribution using MCMC Chain Output
}
\details{
This is a generic function.  See specific method definitions for
additional information.
}
\seealso{
\code{\link{param.post.predict.dynsbm.mcmc}}
}
