% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{plot.dynsbm}
\alias{plot.dynsbm}
\title{Plotting Method for DynSBM Fits}
\usage{
\method{plot}{dynsbm}(x, marginal = TRUE, node.order = c("default",
  "membership"), pal = grey((50:1)/50), ...)
}
\arguments{
\item{x}{object of class dynsbm}

\item{node.order}{determines ordering of nodes.  "default" uses the original
order of the adjacency matrix.  "membership" first orders the data using the
estimated block membership vector.}

\item{pal}{pallete of colors to use ordered from low to high intensity}

\item{xaxt}{a character which specifies the axis type}

\item{yaxt}{a character which specifies the axis type}
}
\description{
Plots the expected mean tie probabilities for a given dynsbm object.
}
\details{
Plots a greyscale image of the adjacency matrix.  Darker colors
indicate larger expected intensities.
}
\examples{

}
\seealso{
\code{\link{dynsbm}}
}

