% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{plot.wsbm}
\alias{plot.wsbm}
\title{Plotting Method for WSBM Fits}
\usage{
\method{plot}{wsbm}(x, node.order = c("default", "membership"),
  pal = grey((50:1)/50), ...)
}
\arguments{
\item{x}{object of class wsbm}

\item{node.order}{determines ordering of nodes.  "default" uses the original
order of the adjacency matrix.  "membership" first orders the data using the
estimated block membership vector.}

\item{pal}{pallete of colors to use ordered from low to high intensity}

\item{xaxt}{a character which specifies the axis type}

\item{yaxt}{a character which specifies the axis type}
}
\description{
Plots the expected mean tie probabilities for a given wsbm object.
}
\details{
Plots a greyscale image of the adjacency matrix.  Darker colors
indicate larger expected intensities.
}
\examples{
data(commuter30)
fit.1 <- wsbm.fit(commuter30,kk=3,hours=1)
plot(fit.1)

}
\seealso{
\code{\link{wsbm.fit}}
}

