% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{post.pred.plot}
\alias{post.pred.plot}
\title{Posterior Predictive Plotting for DynSBM Parameters}
\usage{
post.pred.plot(object, tclass, param = c("block", "sender", "receiver"), ...)
}
\arguments{
\item{object}{object of class "dynsbm"}

\item{tclass}{equivalence class for which to plot posterior predictive
distribution}

\item{param}{which parameters to use.  The default value "block"
plots the block intensity parameters.  Passing "sender" and "receiver" plots
the chains for the sender and receiver effects.}

\item{...}{other graphical parameters to pass to the plotting functions}
}
\description{
Posterior Predictive Plotting for DynSBM Parameters
}
