% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{post.pred.test}
\alias{post.pred.test}
\title{Posterior Predictive Plotting for DynSBM Parameters}
\usage{
post.pred.test(object.1, object.2, param = c("block", "sender", "receiver",
  "all"), ...)
}
\arguments{
\item{object.1}{base object of class "dynsbm.mcmc"}

\item{object.2}{test object of class "dynsbm.mcmc"}

\item{param}{which parameters to use.  The default value "block"
plots the block intensity parameters.  Passing "sender" and "receiver" plots
the chains for the sender and receiver effects.}

\item{...}{other graphical parameters to pass to the plotting functions}
}
\value{
Returns result of a posterior predictive test on the networks in
  object.2 for the given parameters.  The test is to determine if the
  parameters for each network in object.2 came from the same distribution
  as the networks in object.1.  It is assumed that all of the networks in
  object.1 come from the same underlying distribution for each equivalence
  class.
}
\description{
Posterior Predictive Plotting for DynSBM Parameters
}
