% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{post.predict.dynsbm.mcmc}
\alias{post.predict.dynsbm.mcmc}
\title{Posterior Predictive Distribution Method for DynSBM Fits}
\usage{
\method{post.predict}{dynsbm.mcmc}(object, marginal = TRUE, ...)
}
\arguments{
\item{object}{a fitted object of the class "dynsbm".  Requires
clean.out = FALSE in the call to dynsbm.}

\item{marginal}{if TRUE, values block, sender, and receiver effects are
marginalized over using the estimated prior parameters.}

\item{...}{additional parameters}
}
\value{
Returns a 3 dimensional array of dimension n x n x total, where n
is the number of nodes in the network and total is the number of draws in
the MCMC chain generated using wsbm.
}
\description{
Generates a posterior predictive distribution for each tie in a network.
}
\seealso{
\code{\link{dynsbm}} \code{\link{predict.dynsbm}}
}
