% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{post.predict.wsbm.mcmc}
\alias{post.predict.wsbm.mcmc}
\title{Posterior Predictive Distribution Method for WSBM Fits}
\usage{
\method{post.predict}{wsbm.mcmc}(x, ...)
}
\arguments{
\item{x}{a fitted object of the class "wsbm".  Requires clean.out = FALSE
in the call to wsbm.}
}
\value{
Returns a 3 dimensional array of dimension n x n x total, where n
is the number of nodes in the network and total is the number of draws in
the MCMC chain generated using wsbm.
}
\description{
Generates a posterior predictive distribution for each tie in a network.
}
\examples{
data(commuter30)
fit.1 <- wsbm.fit(commuter30,kk=3,hours=1)

pp.dist <- post.pred(fit.1)
density.ppd <- apply(post.pred,3,mean)
summary(density.ppd)


}
\seealso{
\code{\link{wsbm.fit}} \code{\link{predict.wsbm}}
}

