% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{predict.dynsbm}
\alias{predict.dynsbm}
\title{Predict Method for DynSBM Fits}
\usage{
\method{predict}{dynsbm}(x, marginal = TRUE, reps = 1000, ...)
}
\arguments{
\item{object}{a fitted object of the class "dynsbm"}
}
\value{
Returns an array of predicted intensities.  These intensities are
the predicted values given the posterior mean parameter estimates.
}
\description{
Generates a matrix of predicted intensities for each node pair
}
\examples{
data(commuter30)
fit.wsbm <- wsbm(commuter30,kk=3,hours=1)
pred.mat <- predict(fit.wsbm)

}
\seealso{
\code{\link{dynsbm}}
}

