% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{predict.wsbm}
\alias{predict.wsbm}
\title{Predict Method for WSBM Fits}
\usage{
\method{predict}{wsbm}(object, ...)
}
\arguments{
\item{object}{a fitted object of the class "wsbm"}
}
\value{
Returns a matrix of predicted intensities.  These intensities are the
predicted values given the posterior mean parameter estimates
}
\description{
Generates a matrix of predicted intensities for each node pair
}
\examples{
data(commuter30)
fit.1 <- wsbm.fit(commuter30,kk=3,hours=1)
pred.mat <- predict(fit.1)

}
\seealso{
\code{\link{wsbm.fit}}, \code{\link{wsbm}}
}

