% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{wsbm.convergence}
\alias{wsbm.convergence}
\title{Autoconvergence Parameters for wsbm}
\usage{
wsbm.convergence(alpha = 0.001, extend.max = 20, shift.size = 100)
}
\arguments{
\item{alpha}{Confidence level for check}

\item{extend.max}{Maximum number of extensions of the chain while
checking for convergence}

\item{shift.size}{Number of additional thinned steps to run before checking
again for convergence.}
}
\value{
Returns a list containing the three parameters.  Default values
are assigned for all missing parameters
}
\description{
Generating parameters for convergence checking of wsbm MCMC chain
}
\seealso{
\code{\link{wsbm}}
}

