% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{wsbm.fit}
\alias{wsbm.fit}
\title{Weighted Stochastic Block Model Estimator}
\usage{
wsbm.fit(net, kk = 3, hours = 1, self.ties = TRUE,
  priors = wsbm.priors(eta = rep(1/kk, kk)),
  init.control = list(spectral.start = FALSE, multistart = 0, multistart.total
  = 20), mcmc.control = list(total = 1000, burn.in = 1000, thin = 10,
  extend.alpha = 0.001, extend.max = 10, label.switch.mode = "adhoc",
  label.switch.max = 200, multi.impute = FALSE), clean.out = FALSE,
  verbose = 1)
}
\arguments{
\item{net}{a matrix object representing the adjacency matrix}

\item{kk}{number of blocks}

\item{hours}{scaling parameter}

\item{self.ties}{if true assumes self ties are possible}

\item{priors}{priors for posterior inference.  See wsbm.priors for more
details}

\item{init.control}{control parameters for initialization routine.
If init.control contains a member named init.vals it is interpreted as a
list of parameters from which to start the MCMC chain.}

\item{mcmc.control}{control parameters for the MCMC algorithm}

\item{clean.out}{if true, removes MCMC chain from output and only
returns posterior means}

\item{verbose}{higher values correspond to more informative output as the
sampler runs}
}
\value{
Returns a wsbm object that has posterior means for each parameter
in the weighted directed degree corrected stochastic block model.  If
clean.out is FALSE, this object also contains an element 'chain' which
contains the thinned and burned in draws from the MCMC sampler.  The prior
parameters from the call to wsbm are also included.
}
\description{
Estimates parameters for the weighted stochastic block model using
an MCMC algorithm.
}
\seealso{
\code{\link{wsbm.priors}}
}

