% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{wsbm.metric}
\alias{wsbm.metric}
\title{Predictive Matrix for WSBM Fits}
\usage{
wsbm.metric(net, kk = 2, hours = 1, self.ties = TRUE,
  priors = wsbm.priors(eta = rep(1/kk, kk)), verbose = 0, ...)
}
\arguments{
\item{net}{a matrix object representing the adjacency matrix}

\item{kk}{number of blocks}

\item{hours}{scaling parameter}

\item{self.ties}{if true assumes self ties are possible}

\item{priors}{priors for posterior inference.  See wsbm.priors for more details}

\item{verbose}{higher values correspond to more informative output as the}

\item{...}{other parameters to be passed to wsbm}
}
\value{
Returns a matrix object corresponding to the posterior mean intensities.
}
\description{
Generates predicted intensities for a WSBM model given a network.
}
\examples{
data(commuter30)
met.3 <- wsbm.metric(commuter30$net,kk=3)

}
\seealso{
\code{\link{wsbm}} \code{\link{predict.wsbm}}
}
