% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{wsbm.priors}
\alias{wsbm.priors}
\title{Prior Parameters for wsbm Gamma Priors}
\usage{
wsbm.priors(block.alpha = 1, block.beta = 1, sender.alpha = 10,
  sender.beta = 10, receiver.alpha = 10, receiver.beta = 10)
}
\arguments{
\item{block.alpha}{shape parameter for BB}

\item{block.beta}{rate parameter for BB}

\item{sender.alpha}{shape parameter for SS}

\item{sender.beta}{rate parameter for SS}

\item{receiver.alpha}{shape parameter for RR}

\item{receiver.beta}{rate parameter for RR}
}
\value{
Returns a list containing the parameters with useful defaults.
}
\description{
Generating prior paramerters for gamma distributions over Poisson rate
parameters
}
\seealso{
\code{\link{wsbm}}
}

