% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{wsbm}
\alias{wsbm}
\title{Weighted Stochastic Block Model Estimator}
\usage{
wsbm(net, kk = 3, hours = 1, self.ties = TRUE, total = 1000,
  burn.in = 500, thin = 10, priors = wsbm.priors(), eta = rep(1/kk, kk),
  init.vals = NULL, spectral.start = FALSE, clean.out = FALSE,
  verbose = 0, autoconverge = wsbm.convergence(), multistart = 0,
  multistart.total = 20, label.switch.mode = c("adhoc", "kl-loss"),
  max.runs = 200, multiImpute = FALSE)
}
\arguments{
\item{net}{Adjacency matrix}

\item{kk}{Number of blocks}

\item{hours}{Scaling parameter}

\item{self.ties}{If true assumes self ties are possible}

\item{total}{Number of draws output by MCMC sampler}

\item{burn.in}{Number of iterations to discard}

\item{thin}{Amount of draws to thin by}

\item{priors}{Priors for MCMC algorithm.  See wsbm.priors for more details}

\item{eta}{Prior on block membership vector}

\item{init.vals}{List containing parameters for initialization}

\item{spectral.start}{If true, a spectral clustering algorithm is used to
obtain initial parameter estimates}

\item{clean.out}{If true, removes MCMC chain from output and only
returns posterior means}

\item{verbose}{Higher values correspond to more informative output as the
sampler runs}

\item{autoconverge}{Parameters for checking convergence of MCMC chain.
See wsbm.convergence for more details.}

\item{multistart}{If multistart is greater than 0, multistart random
initializations are used to find an optimal starting point for the
MCMC algorithm}

\item{multistart.total}{Number of iterations to perform for each
multistart chain}

\item{label.switch.mode}{Methods for correcting label switching issues
within MCMC chain.  The "kl-loss" method is more principled but very slow.}
}
\value{
Returns a wsbm object that has posterior means for each parameter
in the weighted directed degree corrected stochastic block model.  If
clean.out is FALSE, this object also contains an element 'chain' which
contains the thinned and burned in draws from the MCMC sampler.  The prior
parameters from the call to wsbm are also included.
}
\description{
Estimates parameters for the weighted stochastic block model using
an MCMC algorithm.
}
\seealso{
\code{\link{wsbm.priors}}, \code{\link{wsbm.convergence}}
}

