% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{AIC.dynsbm}
\alias{AIC.dynsbm}
\title{AIC method for DynSBM Fits}
\usage{
\method{AIC}{dynsbm}(object, ..., k = 2)
}
\arguments{
\item{object}{an object of class "dynsbm"}

\item{...}{additional arguments to AIC}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2
is the classical AIC}
}
\value{
returns the Akaike Information Criterion using the posterior mean
parameters as approximations to the MLE.  The number of parameters is taken
to be T*(2*(n-k) + k*k) + n.
}
\description{
Returns Akaike Information Criterion for DynSBM Fit
}
\seealso{
\code{\link{dynsbm}}
}
