% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{diagnostic.plot}
\alias{diagnostic.plot}
\title{Diagnostic Plotting Function}
\usage{
diagnostic.plot(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{additional arguments for specific methods}
}
\description{
Plots Diagnostics for Various Model Fits
}
\details{
This is a generic funciton.  See specific method definitions for
additional information.
}
\seealso{
\code{\link{diagnostic.plot.wsbm.mcmc}},
\code{\link{diagnostic.plot.dynsbm.mcmc}}
}

