% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{DIC.dynsbm}
\alias{DIC.dynsbm}
\title{DIC method for DynSBM Fits}
\usage{
\method{DIC}{dynsbm}(object, ...)
}
\arguments{
\item{object}{an object of class "dynsbm"}

\item{...}{additional parameters}
}
\value{
returns the Deviance Information Criterion.
}
\description{
Returns Deviance Information Criterion for DynSBM Fit
}
\seealso{
\code{\link{dynsbm}}
}
