% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{dynsbm.priors}
\alias{dynsbm.priors}
\title{Prior Parameters for dynsbm}
\usage{
dynsbm.priors(eta, BB.hyperprior = c(0.99, 1.01, 1, 1),
  SS.hyperprior = c(0.99, 1.01, 1, 1), RR.hyperprior = c(0.99, 1.01, 1, 1))
}
\arguments{
\item{eta}{prior on block memberships}

\item{block.alpha}{shape parameter for Block Effects}

\item{block.beta}{rate parameter for Block Effects}

\item{sender.alpha}{shape parameter for Sender Effects}

\item{sender.beta}{rate parameter for Sender Effects}

\item{receiver.alpha}{shape parameter for Receiver Effects}

\item{receiver.beta}{rate parameter for Receiver Effects}
}
\value{
Returns a list containing the parameters with useful defaults.  eta
is the only required parameter.
}
\description{
Generating prior parameters for pass to dynsbm
}
\seealso{
\code{\link{dynsbm}}
}

