% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/wsbm.R, R/dynamic-sbm.R
\name{get.iter}
\alias{get.iter}
\alias{get.iter.wsbm.mcmc}
\alias{get.iter.dynsbm.mcmc}
\title{Get Model Object for an MCMC Iteration}
\usage{
get.iter(object, iter, ...)

\method{get.iter}{wsbm.mcmc}(object, iter, ...)

\method{get.iter}{dynsbm.mcmc}(object, iter, ...)
}
\arguments{
\item{object}{mdl.mcmc object}

\item{iter}{iteration to return}

\item{...}{additional parameters}
}
\value{
Returns an object of the class for which the given MCMC object
  is designed to sample.
}
\description{
Returns Model Object from MCMC Iteration
}
\details{
Assuming the MCMC sampler is iteratively drawing parameters from
  the joint distribution for a given model class, this function will
  return the sample from a given iteration of the correct class.  This
  model object can then be used as a starting point for extending a chain,
  or can be analyzed by any of the functions that allow you to interact
  with that model object.
}
\section{Methods (by class)}{
\itemize{
\item \code{wsbm.mcmc}: returns "wsbm" model object for a specific iteration.

\item \code{dynsbm.mcmc}: returns "dynsbm" model object for a specific iteration.
}}

\seealso{
\code{\link{wsbm.fit}}, \code{\link{dynsbm.fit}}
}
