% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{logLik.wsbm}
\alias{logLik.wsbm}
\title{Log-Likelihood method for WSBM Fits}
\usage{
\method{logLik}{wsbm}(x)
}
\arguments{
\item{x}{an object of class "wsbm"}
}
\value{
returns the log-likelihood of the data given the posterior mean
parameter estimates.
}
\description{
Returns log-likelihood for posterior mean estimates.
}
\seealso{
\code{\link{wsbm}}
}

