% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{net.gen.dynsbm}
\alias{net.gen.dynsbm}
\title{Generate Set of Dynamic Stochastic Block Model Networks}
\usage{
\method{net.gen}{dynsbm}(object, marginal = FALSE, ...)
}
\arguments{
\item{object}{object of class "dynsbm"}

\item{marginal}{if TRUE, parameters are sampled from prior distributions
before simulating a new network}

\item{...}{additional parameters}
}
\value{
Returns an n x n x T array representing a collection of T
networks with n nodes corresponding to the dynsbm object.
}
\description{
Generates networks from the weighted stochastic block model
with or without degree effects.
}
\examples{
kk <- 3
nn <- kk^2 * 10
TT <- 10; ee <- 2
tmap <- rep(1:ee,TT/ee)
hours.vec <- c(12,12)
mmb <-  rep(1:kk,each=nn/kk)

###  Setting up Reasonable Prior Distributions
BB.prior <- array(1,c(kk,kk,2,ee))
BB.prior[cbind(1:3,c(2,3,1),1,1)] <- 2
BB.prior[cbind(1:3,c(3,1,2),1,2)] <- 2

SS.prior <- array(rgamma(nn*2*ee,10,10),c(nn,2,ee))
RR.prior <- array(rgamma(nn*2*ee,10,10),c(nn,2,ee))

###  Setting up Object with Dynsbm Parameters
dat <- dynsbm(nn=nn,TT=TT,tmap=tmap,hours.vec=hours.vec,
              BB.prior=BB.prior,SS.prior=SS.prior,RR.prior=RR.prior,
              mmb=mmb,self.ties=TRUE,normalize=TRUE)

###  Sampling New Edges Only
net.mat.1 <- net.gen(dat)
network.plot(net.mat.1)

net.mat.2 <- net.gen(dat)
network.plot(net.mat.2)

net.mat.3 <- net.gen(dat)
network.plot(net.mat.3)

###  Resampling Parameters from Priors and Edges
net.mat.marg.1 <- net.gen(dat,marginal=TRUE)
network.plot(net.mat.marg.1)

net.mat.marg.2 <- net.gen(dat,marginal=TRUE)
network.plot(net.mat.marg.2)

net.mat.marg.3 <- net.gen(dat,marginal=TRUE)
network.plot(net.mat.marg.3)

}
\seealso{
\code{\link{dynsbm}}, \code{\link{dynsbm.fit}}
}
