% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/helper.R, R/wsbm.R,
%   R/dynamic-sbm.R
\name{network.plot}
\alias{network.plot}
\alias{network.plot.matrix}
\alias{network.plot.array}
\alias{network.plot.wsbm}
\alias{network.plot.dynsbm}
\title{Network Plot}
\usage{
network.plot(x, ...)

\method{network.plot}{matrix}(x, pal = grey((50:1)/50), node.order = NULL,
  ...)

\method{network.plot}{array}(x, pal = grey((50:1)/50), node.order = NULL,
  ...)

\method{network.plot}{wsbm}(x, pal = grey((50:1)/50), node.order = NULL,
  ...)

\method{network.plot}{dynsbm}(x, pal = grey((50:1)/50), node.order = NULL,
  ...)
}
\arguments{
\item{x}{network object}

\item{...}{additional graphical parameters}

\item{pal}{Color palette used to indicate tie intensity}

\item{node.order}{order in which to plot the nodes}
}
\description{
Plots a Heatmap Representation of an Adjacency Matrix
}
\details{
Plots a heatmap representation of the adjacency matrix for a
network object.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Plots heatmap assuming x is an adjacency matrix

\item \code{array}: Plots heatmap for multiple adjacency matrices
combined into a single 3-dimensional array.  Dimensions 1 and 2 are
assumed to correspond to the nodes, and dimension 3 the different
networks.

\item \code{wsbm}: Plots heatmap for net

\item \code{dynsbm}: Plots heatmap for each network in net.mat
}}

\seealso{
\code{\link{wsbm}},
\code{\link{dynsbm}}
}
