% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param.plot}
\alias{param.plot}
\title{Parameter Plotting Function}
\usage{
param.plot(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{additional arguments for specific methods}
}
\description{
Plots Parameter Summaries for Various Model Fits
}
\details{
This is a generic funciton.  See specific method definitions for
additional information.
}
\seealso{
\code{\link{param.plot.wsbm.mcmc}}
\code{\link{param.plot.dynsbm.mcmc}}
}

