% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{plot.dynsbm}
\alias{plot.dynsbm}
\title{Plotting Method for DynSBM Fits}
\usage{
\method{plot}{dynsbm}(x, marginal = TRUE, node.order = c("default",
  "membership"), pal = grey((50:1)/50), ...)
}
\arguments{
\item{x}{object of class dynsbm}

\item{marginal}{if TRUE average tie probabilities for the equivalence
classes are plotted.  If marginal is FALSE, the average tie probability
for each individual network is plotted instead.}

\item{node.order}{determines ordering of nodes.  "default" uses the original
order of the adjacency matrix.  "membership" first orders the data using the
estimated block membership vector.}

\item{pal}{color palette used to indicate tie intensity}

\item{...}{additional graphical parameters}
}
\description{
Plots the expected mean tie probabilities for a given dynsbm object.
}
\details{
Plots a greyscale image of the mean tie intensity matrix.  The
  individual network parameters are marginalized over each equivalence
  class when marginal = TRUE.
}
\seealso{
\code{\link{dynsbm}}
}
