% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-sbm.R
\name{predict.dynsbm}
\alias{predict.dynsbm}
\title{Predict Method for DynSBM Fits}
\usage{
\method{predict}{dynsbm}(object, marginal = TRUE, reps = 1000, ...)
}
\arguments{
\item{object}{a fitted object of the class "dynsbm"}

\item{marginal}{if TRUE, prediction is made using the marginal model.
Otherwise, the model conditional on parameter estimates is used}

\item{reps}{Number of samples from the marginal model to use when marginal
= TRUE}

\item{...}{additional parameters}
}
\value{
Returns an array of predicted intensities.  These intensities are
the predicted values given the posterior mean parameter estimates.
}
\description{
Generates a matrix of predicted intensities for each node pair
}
\seealso{
\code{\link{dynsbm.fit}}
}
