% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{wsbm.priors}
\alias{wsbm.priors}
\title{Prior Parameters for wsbm Gamma Priors}
\usage{
wsbm.priors(eta, block.alpha = 1, block.beta = 1, sender.alpha = 10,
  sender.beta = 10, receiver.alpha = 10, receiver.beta = 10)
}
\arguments{
\item{eta}{prior on block memberships}

\item{block.alpha}{shape parameter for Block Effects}

\item{block.beta}{rate parameter for Block Effects}

\item{sender.alpha}{shape parameter for Sender Effects}

\item{sender.beta}{rate parameter for Sender Effects}

\item{receiver.alpha}{shape parameter for Receiver Effects}

\item{receiver.beta}{rate parameter for Receiver Effects}
}
\value{
Returns a list containing the parameters with useful defaults.  eta
is the only required parameter.
}
\description{
Generating prior paramerters for gamma distributions over Poisson rate
parameters
}
\seealso{
\code{\link{wsbm}}
}
