% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbm.R
\name{wsbm}
\alias{wsbm}
\title{Create Weighted Stochastic Block Model Object}
\usage{
wsbm(nn, mmb, BB, SS = rep(1, nn), RR = rep(1, nn), mmb.prior = c(0.5,
  0.5), self.ties = TRUE, hours = 1, gen = FALSE)
}
\arguments{
\item{nn}{number of nodes in the network}

\item{mmb}{block memberhsip vector}

\item{BB}{block propensity matrix}

\item{SS}{sender effect vector}

\item{RR}{receiver effect vector}

\item{mmb.prior}{prior distribution for multinomial distribution over
block memberships}

\item{self.ties}{if true self ties are allowed}

\item{hours}{scaling factor so that other factors can be interpreted
as hourly rate parameters.}

\item{gen}{if TRUE also samples a network from the model and includes it as a
component of the model object}
}
\value{
Returns a "wsbm" model object containing all the parameters needed
for sampling networks from the model.  If gen = TRUE, the object also
contains a network sampled from this parameterization.
}
\description{
Generates a Model Object for a Weighted Stochastic Block Model
}
\examples{
kk <- 3
nn <- kk^2 * 10
mmb <-  rep(1:kk,each=nn/kk)
BB <- array(1,c(kk,kk)); diag(BB) <- 5
SS <- RR <- rep(c(rep(5,nn/(kk^2)),rep(1,nn*(kk-1)/(kk^2))),kk)
set.seed(100)
dat <- wsbm(nn=nn,mmb=mmb,BB=BB,SS=SS,RR=RR,gen=TRUE)

### Plotting Intensity Matrix and Adjacency Matrix
par(mfrow=c(1,2))
plot(dat,main="Mean Intensity Matrix")
network.plot(dat,main="Adjacency Matrix")

}
\seealso{
\code{\link{wsbm.fit}}
}

